--- a/packages/docusaurus-utils/package.json
+++ b/packages/docusaurus-utils/package.json
@@ -33,6 +33,7 @@
     "lodash": "^4.17.21",
     "micromatch": "^4.0.5",
     "prompts": "^2.4.2",
+    "p-queue": "^6.6.2",
     "resolve-pathname": "^3.0.0",
     "tslib": "^2.6.0",
     "url-loader": "^4.1.1",