--- a/packages/docusaurus/src/commands/build/buildLocale.ts
+++ b/packages/docusaurus/src/commands/build/buildLocale.ts
@@ -58,10 +58,6 @@ export async function buildLocale({
     }),
   );
 
-  if (process && site) {
-    process.exit(0);
-  }
-
   const {props} = site;
   const {outDir, plugins, siteConfig} = props;