## Motivation


Fix https://github.com/facebook/docusaurus/issues/10348

Avoids Node.js EBADF error on large Docusaurus sites due to running too many concurrent Git commands.

I choose `os.availableParallelism() * 4;` as an arbitrary default concurrency value. It's high on purpose to be conservative with the former behavior, which was Infinity. As my local benchmark shows, between 4 and Infinity, the performance impact is quite low anyway.

## Benchmark

It's hard to provide the perfect default value, but we can see on my Mac M3 Pro that:
- Concurrency 1 is the slowest
- Concurrency 0 (= Infinity) is slower than any concurrency > 4
- On my computer (12 cores), the sweet spot seems to be in between concurrency 10 -> 100

```bash
hyperfine --runs 4 \
'DOCUSAURUS_GIT_COMMAND_CONCURRENCY=0 yarn build:website --locale en' \
'DOCUSAURUS_GIT_COMMAND_CONCURRENCY=1 yarn build:website --locale en' \
'DOCUSAURUS_GIT_COMMAND_CONCURRENCY=2 yarn build:website --locale en' \
'DOCUSAURUS_GIT_COMMAND_CONCURRENCY=4 yarn build:website --locale en' \
'DOCUSAURUS_GIT_COMMAND_CONCURRENCY=8 yarn build:website --locale en' \
'DOCUSAURUS_GIT_COMMAND_CONCURRENCY=12 yarn build:website --locale en' \
'DOCUSAURUS_GIT_COMMAND_CONCURRENCY=16 yarn build:website --locale en' \
'DOCUSAURUS_GIT_COMMAND_CONCURRENCY=20 yarn build:website --locale en' \
'DOCUSAURUS_GIT_COMMAND_CONCURRENCY=24 yarn build:website --locale en' \
'DOCUSAURUS_GIT_COMMAND_CONCURRENCY=30 yarn build:website --locale en' \
'DOCUSAURUS_GIT_COMMAND_CONCURRENCY=40 yarn build:website --locale en' \
'DOCUSAURUS_GIT_COMMAND_CONCURRENCY=60 yarn build:website --locale en' \
'DOCUSAURUS_GIT_COMMAND_CONCURRENCY=100 yarn build:website --locale en'

Benchmark 1: DOCUSAURUS_GIT_COMMAND_CONCURRENCY=0 yarn build:website --locale en
  Time (mean ± σ):      4.255 s ±  1.074 s    [User: 12.487 s, System: 6.392 s]
  Range (min … max):    3.703 s …  5.866 s    4 runs
 
Benchmark 2: DOCUSAURUS_GIT_COMMAND_CONCURRENCY=1 yarn build:website --locale en
  Time (mean ± σ):     13.151 s ±  0.083 s    [User: 9.875 s, System: 3.586 s]
  Range (min … max):   13.061 s … 13.246 s    4 runs
 
Benchmark 3: DOCUSAURUS_GIT_COMMAND_CONCURRENCY=2 yarn build:website --locale en
  Time (mean ± σ):      7.614 s ±  0.023 s    [User: 9.894 s, System: 3.567 s]
  Range (min … max):    7.581 s …  7.633 s    4 runs
 
Benchmark 4: DOCUSAURUS_GIT_COMMAND_CONCURRENCY=4 yarn build:website --locale en
  Time (mean ± σ):      4.823 s ±  0.008 s    [User: 10.097 s, System: 3.688 s]
  Range (min … max):    4.817 s …  4.832 s    4 runs
 
Benchmark 5: DOCUSAURUS_GIT_COMMAND_CONCURRENCY=8 yarn build:website --locale en
  Time (mean ± σ):      3.698 s ±  0.016 s    [User: 11.257 s, System: 4.917 s]
  Range (min … max):    3.677 s …  3.716 s    4 runs
 
Benchmark 6: DOCUSAURUS_GIT_COMMAND_CONCURRENCY=12 yarn build:website --locale en
  Time (mean ± σ):      3.617 s ±  0.015 s    [User: 11.828 s, System: 5.437 s]
  Range (min … max):    3.601 s …  3.635 s    4 runs
 
Benchmark 7: DOCUSAURUS_GIT_COMMAND_CONCURRENCY=16 yarn build:website --locale en
  Time (mean ± σ):      3.615 s ±  0.024 s    [User: 11.915 s, System: 5.495 s]
  Range (min … max):    3.588 s …  3.647 s    4 runs
 
Benchmark 8: DOCUSAURUS_GIT_COMMAND_CONCURRENCY=20 yarn build:website --locale en
  Time (mean ± σ):      3.617 s ±  0.042 s    [User: 11.822 s, System: 5.472 s]
  Range (min … max):    3.579 s …  3.671 s    4 runs
 
Benchmark 9: DOCUSAURUS_GIT_COMMAND_CONCURRENCY=24 yarn build:website --locale en
  Time (mean ± σ):      3.600 s ±  0.007 s    [User: 11.873 s, System: 5.512 s]
  Range (min … max):    3.589 s …  3.604 s    4 runs
 
Benchmark 10: DOCUSAURUS_GIT_COMMAND_CONCURRENCY=30 yarn build:website --locale en
  Time (mean ± σ):      3.613 s ±  0.027 s    [User: 11.841 s, System: 5.496 s]
  Range (min … max):    3.589 s …  3.649 s    4 runs
 
Benchmark 11: DOCUSAURUS_GIT_COMMAND_CONCURRENCY=40 yarn build:website --locale en
  Time (mean ± σ):      3.623 s ±  0.012 s    [User: 11.872 s, System: 5.503 s]
  Range (min … max):    3.613 s …  3.639 s    4 runs
 
Benchmark 12: DOCUSAURUS_GIT_COMMAND_CONCURRENCY=60 yarn build:website --locale en
  Time (mean ± σ):      3.644 s ±  0.039 s    [User: 11.948 s, System: 5.574 s]
  Range (min … max):    3.616 s …  3.700 s    4 runs
 
Benchmark 13: DOCUSAURUS_GIT_COMMAND_CONCURRENCY=100 yarn build:website --locale en
  Time (mean ± σ):      3.650 s ±  0.027 s    [User: 11.985 s, System: 5.563 s]
  Range (min … max):    3.610 s …  3.669 s    4 runs
 
Summary
  DOCUSAURUS_GIT_COMMAND_CONCURRENCY=24 yarn build:website --locale en ran
    1.00 ± 0.01 times faster than DOCUSAURUS_GIT_COMMAND_CONCURRENCY=30 yarn build:website --locale en
    1.00 ± 0.01 times faster than DOCUSAURUS_GIT_COMMAND_CONCURRENCY=16 yarn build:website --locale en
    1.00 ± 0.00 times faster than DOCUSAURUS_GIT_COMMAND_CONCURRENCY=12 yarn build:website --locale en
    1.00 ± 0.01 times faster than DOCUSAURUS_GIT_COMMAND_CONCURRENCY=20 yarn build:website --locale en
    1.01 ± 0.00 times faster than DOCUSAURUS_GIT_COMMAND_CONCURRENCY=40 yarn build:website --locale en
    1.01 ± 0.01 times faster than DOCUSAURUS_GIT_COMMAND_CONCURRENCY=60 yarn build:website --locale en
    1.01 ± 0.01 times faster than DOCUSAURUS_GIT_COMMAND_CONCURRENCY=100 yarn build:website --locale en
    1.03 ± 0.00 times faster than DOCUSAURUS_GIT_COMMAND_CONCURRENCY=8 yarn build:website --locale en
    1.18 ± 0.30 times faster than DOCUSAURUS_GIT_COMMAND_CONCURRENCY=0 yarn build:website --locale en
    1.34 ± 0.00 times faster than DOCUSAURUS_GIT_COMMAND_CONCURRENCY=4 yarn build:website --locale en
    2.12 ± 0.01 times faster than DOCUSAURUS_GIT_COMMAND_CONCURRENCY=2 yarn build:website --locale en
    3.65 ± 0.02 times faster than DOCUSAURUS_GIT_COMMAND_CONCURRENCY=1 yarn build:website --locale en
```